Avoid a signed-unsigned comparison warning
PiperOrigin-RevId: 209166600

+++ asylo/platform/common/bridge_functions.cc
@@ -752,7 +752,7 @@ int BridgeCpuSetCheckBit(int cpu, BridgeCpuSet *set) {
 
 bool CStringCopy(const char *source_buf, char *dest_buf, size_t size) {
   int ret = snprintf(dest_buf, size, "%s", source_buf);
-  return ret >= 0 && ret < size;
+  return ret >= 0 && static_cast<size_t>(ret) < size;
 }
 
 }  // namespace asylo
